refactor: router and file structure
