+++ src/App.tsx
@@ -2,13 +2,13 @@ import React from 'react'
 import { Provider } from 'react-redux'
 import './App.css'
 import { store } from './app/store'
-import Home from './components/pages/Home'
+import Router from './Router'
 
 function App() {
   return (
     <Provider store={store}>
       <div className="App">
-        <Home />
+        <Router />
       </div>
     </Provider>
   )
