+++ src/components/shared/ArticleSection.tsx
