+++ src/components/shared/ArticleThumbnail.test.js
