+++ src/components/shared/ArticleThumbnail.tsx
